refactor from ant to maven, all unit tests passed except the Btree in lab4,5,6

+++ lab5/src/main/java/simpledb/OperatorCardinality.java
@@ -121,8 +121,8 @@ public class OperatorCardinality {
             child1Card = child1O.getEstimatedCardinality();
             child1Card = child1Card > 0 ? child1Card : 1;
         } else if (child1 instanceof SeqScan) {
-            child1Card = (int) (tableStats.get(((SeqScan) child1)
-                    .getTableName()).estimateTableCardinality(1.0));
+            child1Card = tableStats.get(((SeqScan) child1)
+                    .getTableName()).estimateTableCardinality(1.0);
         }
 
         if (child2 instanceof Operator) {
@@ -133,8 +133,8 @@ public class OperatorCardinality {
             child2Card = child2O.getEstimatedCardinality();
             child2Card = child2Card > 0 ? child2Card : 1;
         } else if (child2 instanceof SeqScan) {
-            child2Card = (int) (tableStats.get(((SeqScan) child2)
-                    .getTableName()).estimateTableCardinality(1.0));
+            child2Card = tableStats.get(((SeqScan) child2)
+                    .getTableName()).estimateTableCardinality(1.0);
         }
 
         j.setEstimatedCardinality(JoinOptimizer.estimateTableJoinCardinality(j
@@ -164,11 +164,9 @@ public class OperatorCardinality {
         boolean child1HasJoinPK = Database.getCatalog()
                 .getPrimaryKey(tableAliasToId.get(tableAlias1))
                 .equals(pureFieldName1);
-        ;
         boolean child2HasJoinPK = Database.getCatalog()
                 .getPrimaryKey(tableAliasToId.get(tableAlias2))
                 .equals(pureFieldName2);
-        ;
 
         if (child1 instanceof Operator) {
             Operator child1O = (Operator) child1;
@@ -178,8 +176,8 @@ public class OperatorCardinality {
             child1Card = child1O.getEstimatedCardinality();
             child1Card = child1Card > 0 ? child1Card : 1;
         } else if (child1 instanceof SeqScan) {
-            child1Card = (int) (tableStats.get(((SeqScan) child1)
-                    .getTableName()).estimateTableCardinality(1.0));
+            child1Card = tableStats.get(((SeqScan) child1)
+                    .getTableName()).estimateTableCardinality(1.0);
         }
 
         if (child2 instanceof Operator) {
@@ -190,8 +188,8 @@ public class OperatorCardinality {
             child2Card = child2O.getEstimatedCardinality();
             child2Card = child2Card > 0 ? child2Card : 1;
         } else if (child2 instanceof SeqScan) {
-            child2Card = (int) (tableStats.get(((SeqScan) child2)
-                    .getTableName()).estimateTableCardinality(1.0));
+            child2Card = tableStats.get(((SeqScan) child2)
+                    .getTableName()).estimateTableCardinality(1.0);
         }
 
         j.setEstimatedCardinality(JoinOptimizer.estimateTableJoinCardinality(j
@@ -220,8 +218,8 @@ public class OperatorCardinality {
         }
 
         if (child instanceof SeqScan) {
-            childCard = (int) (tableStats.get(((SeqScan) child).getTableName())
-                    .estimateTableCardinality(1.0));
+            childCard = tableStats.get(((SeqScan) child).getTableName())
+                    .estimateTableCardinality(1.0);
         }
 
         String[] tmp = a.groupFieldName().split("[.]");
